smart-ingestor/requirements.txt
flask
pandas
httpx
tqdm
backoff
kafka-python